Resolve image name to ID before fetching digest to resolve false drift
EL9.5 ships podman v5 with skopeo 1.16 up from 1.14. The behaviour of how
image digests are reported has changed, causing `podman::container` to
consider needing to fetch a new image on every run and unwanted container
restarts.

`podman image inspect --format='{{.Digest}}' now returns different hashes
depending on whether an image ID or a tag is given as argument.

Skopeo reports the image digest in the registry as if podman image inspect
were called with the image ID.

This code resolves the given image name (which might be an ID or a Tag)
into the ID before trying to fetch the digest. This prevents the false
drift.